support external TIME_PRINT value with --TIME
TIME_PRINT controls the verbosity of time.sh - which shcmp uses
by default when --TIME is used. If it's unset then time.sh sets
it to 3 (verbose).

Till now, shcmp set it to 0 such that it's less noisy, but in a way
which the user cannot override with TIME_PRINT=3 shcmp...

Now it can be overridden in this way and affects the verbosity of
time.sh as expected.