output: always append "(exit:N)"
The explicit exit code is useful, especially on errors, and doesn't
add meaningful "noise" even on success.

This makes the error/ok prefix redundant at cluster_outputs, and
similarly --NOTRIM is no longer required, but for clarity this commit
only changes the output, and the next commit will adapt the options
and other changes.