remove --NOTRIM, update help/readme
Following the previous commit, --NOTRIM is now not required because
the output is always without trailing newlines, and similarly the
ok/error prefix is not required at cluster_outputs.

Make the relevant changes at the code, and update the help/readme text
to reflect this and the previous (global exit code) changes.